canchas_django_PORTERMINAL/requirements.txt
Django==4.2.*
django-environ>=0.11
mysqlclient>=2.2

